Fixed test errors

+++ tests/sem/statement/underscore_usage_4.go
@@ -2,7 +2,7 @@
 //
 package main
 
-func alpha (a int) (int) {
+func alpha (a int) (float64) {
 	return 6.7
 }
